fix(api): Update shape of preprocessed response
Include more information on response, instead of just the body info.
To properly handle errors with fetch & redux-thunk, am going to need to
decide it's an error from inside of the thunk itself. Cannot seem to
reject the promise, throw an error from the preProcess helper.

This will break action creators, but it's only a matter of changing the
key to look to.