fix(state/actions): Catch errors & handle updated response shapes
Need to decide that something is an error from the thunk. Normally,
fetch will only throw when there's an HTTP error, but that should be
overridden when you throw an error or reject the promise, thereby,
bypassing the "then" block entirely, and going straight to "catch".
However, that's not happening, not sure the reason. So, that's why
we're throwing an error from the thunk in order to go into "catch".