feat(state): Store message visibility in state
Was getting annoying error message about the 'dismiss' icon having the
wrong name. It's a bug in Semantic-UI framework. It may or may not have
been fixed. But upgrading semantic-UI is a pain. Am opting to store
visibility in state, so that it'll wipe itself away when needed.
Drawback is, it takes control away from user; they can no longer
manually dismiss the message.